Add files via upload
# Mental-Health-Analyzer

This project focuses on analyzing various factors that could lead to mental health disorders. It uses machine learning techniques to predict which mental health disorder an individual is likely to have based on a set of behavioral and psychological features.

## Overview

Mental health issues are a growing concern, and early detection is critical for providing timely support. This project aims to predict the possibility of mental health disorders based on various symptoms and behaviors. We analyze a dataset with multiple features related to anxiety, depression, stress and loneliness, and apply Logistic Regression to classify individuals.

## Dataset

The dataset includes several features that can predict mental and emotional states, behaviors, and physical symptoms:

- feeling.nervous: Frequency of feeling nervous
- panic: Occurrence of panic attacks
- breathing.rapidly: Instances of rapid breathing
- sweating: Frequency of sweating due to anxiety
- trouble.in.concentration: Difficulty in maintaining concentration
- having.trouble.in.sleeping: Difficulty in sleeping
- having.trouble.with.work: Trouble managing work or tasks
- hopelessness: Feeling of hopelessness
- anger: Experience of anger and frustration
- over.react: Tendency to overreact
- weight.gain: Weight fluctuations
- material.possessions: Attachment to material possessions
- introvert: Introversion tendencies
- popping.up.stressful.memory: Recurrence of stressful memories
- having.nightmares: Frequent nightmares
- avoids.people.or.activities: Avoidance of people or social activities
- feeling.negative: Consistent negative feelings
- trouble.concentrating: Difficulty concentrating
- blaming.yourself: Tendency to blame oneself
- Disorder: Target variable, indicating which mental health disorder an individual have